feat: 인용메타데이터 가져오기

+++ src/main/java/com/phraiz/back/member/domain/Member.java
@@ -23,7 +23,8 @@ public class Member {
     private Long memberId; // 멤버 식별자
 
     @Column(nullable = false)
-    private Long planId;
+    private int planId;
+    // 기본값 0 : free?
 
     @Column(nullable = false, unique = true)
     private String id;
